chore: Refactor latency command code
